refactor: centralize error handling for better-auth API calls in ApiKeysRepository
- Introduced `rethrowAsDomainError` to map `APIError` to meaningful domain errors (`NotFoundError`, `ForbiddenError`, `ValueError`).
- Removed redundant service-level error handling; delegated error handling to the repository layer for consistency.
- Simplified `ApiKeysService` methods by directly invoking repository methods without try-catch blocks.
